move timeout stuff to lapitt, where it is used
