+++ ledger-test-tool/infrastructure/src/main/scala/com/daml/ledger/api/testtool/infrastructure/WithTimeout.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.testing
+package com.daml.ledger.api.testtool.infrastructure
 
 import java.util.Timer
