+++ ledger-test-tool/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/CommandSubmissionCompletionIT.scala
@@ -8,10 +8,9 @@ import java.util.regex.Pattern
 import com.daml.error.definitions.LedgerApiErrors
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
-import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
+import com.daml.ledger.api.testtool.infrastructure.{LedgerTestSuite, TimeoutException, WithTimeout}
 import com.daml.ledger.test.model.Test.Dummy
 import com.daml.ledger.test.model.Test.Dummy._
-import com.daml.platform.testing.{TimeoutException, WithTimeout}
 
 import scala.concurrent.duration.DurationInt
